Migrate Taggable from JVM to common
