+++ kotlinpoet/src/nonJvmMain/kotlin/com/squareup/kotlinpoet/jvm/alias/JvmClass.nonJvm.kt
@@ -1,0 +1,26 @@
+/*
+ * Copyright (C) 2024 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.squareup.kotlinpoet.jvm.alias
+
+import kotlin.reflect.KClass
+
+@JvmTypeAliasKotlinPoetApi
+public actual class JvmClass<@Suppress("unused")
+  T : Any,> private constructor()
+
+@JvmTypeAliasKotlinPoetApi
+public actual val <T : Any> JvmClass<T>.kotlin: KClass<T>
+  get() = throw UnsupportedOperationException()
